compare: add a distance based on convolution.
* Images are padded out to height and width are power of two (blocksize)

* The two images to compare are subtracted and normalized (+1.0 for
  added pixels, -1.0 for removed pixels)

* The image is repeatedly convolved using a 2x2 averaging kernel,
  yielding a sequence of images, each half the size of the previous
  one.

* For each pixel (x,y), the distance is

  sum_{r=1..blocksize} |img_r (x/2^r, y/2^r)|

If a symbol is moved in an image, the effect averages out as soon as
the convolution kernel encompasses the movement.

This means that the image-diff distance is monotone in how much the
original is offset. In other words, a small shift has a smaller effect
than a larger shift.

This doesn't quite work: because we don't don't do a full convolution,
the two pixels may be adjacent but only average out at the highest
level (eg 2 pixels near the middle of the picture.)

A better approach would be 

  sum_{r=1..blocksize} |convolve_{2^r}(x, y)|

where convolve_{X} convolves with a box kernel of size X.